+++ FoodInEye_app/app/src/main/java/com/example/foodineye_app/activity/OrderAdapter.java
@@ -11,6 +11,7 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.foodineye_app.R;
+import com.example.foodineye_app.data.Order;
 
 import java.util.List;
